requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
streamlit
langchain_community
langchain_google_genai
langchain
chromadb
sentence-transformers
sqlalchemy
pydantic
fastapi-middleware
fastapi-middleware.cors
PyPDF2
pdfminer.six
python-docx
pandas